search: store grapheme positions as u32 in State
State is copied into the queue on every enqueue, so its size drives memory
traffic and cache footprint in the BFS. The three grapheme-position fields
(j, matched_start, matched_end) were usize; make them u32, shrinking State
from 48 to 32 bytes and the dedup key from 36 to 24. Positions are grapheme
indices into the haystack, so this caps usable haystacks at ~4 billion
graphemes (>=4 GiB) — far beyond where this fuzzy search is practical.

~2-4% on the search paths, no regressions.

+++ src/structs.rs
@@ -129,13 +129,14 @@ pub type FxHashMap<K, V> = HashMap<K, V, BuildHasherDefault<FxHasher>>;
 /// Edit count type - u8 is sufficient for practical edit distances (max 255)
 pub type NumEdits = u8;
 
-/// Internal search state
+/// Internal search state. The grapheme positions are `u32` (haystacks far below 4 GiB) to keep the
+/// struct — copied on every enqueue — compact and cache-dense.
 #[derive(Clone)]
 pub(crate) struct State {
     pub(crate) node: usize,
-    pub(crate) j: usize,
-    pub(crate) matched_start: usize,
-    pub(crate) matched_end: usize,
+    pub(crate) j: u32,
+    pub(crate) matched_start: u32,
+    pub(crate) matched_end: u32,
     pub(crate) penalties: f32,
     pub(crate) edits: NumEdits,
     pub(crate) insertions: NumEdits,
